Better fix for deleting jobs without a db
The previous fix with postgres still caused problems because it
left a dangling transaction.

+++ rvbd-portal/rvbd_portal/apps/report/management/commands/clean.py
@@ -13,7 +13,7 @@ import optparse
 
 from django.core.management.base import BaseCommand
 from django.core import management
-from django.db import DatabaseError
+from django.db import DatabaseError, connection
 from django.db.models import get_app, get_models, Count
 
 from django.conf import settings
@@ -49,14 +49,13 @@ class Command(BaseCommand):
     )
 
     def handle(self, *args, **options):
+        db_exists = Job._meta.db_table in connection.introspection.table_names()
+
         if options['clear_cache']:
             # first delete all jobs
             self.stdout.write('Clearing all jobs ... ', ending='')
-            try:
+            if db_exists:
                 Job.objects.all().delete()
-            except DatabaseError:
-                # on new db we don't have Job models yet
-                pass
             self.stdout.write('done.')
 
             # now clear any remaining cache files
